Update preconditions.md
The IfBody method is no longer available, I noticed this playing around with the library. It is a handy function, just a question what was the motivation for removing this? Possible to reintroduce it? I like the idea of being able to use the following for my posts 
stub.Post("/api/path/action")
.IfBody(s => s.ReadAsString().Contains("foo") && !s.ReadAsString().Contains("bar"));
.Response((request, args) => "foo response");

stub.Post("/api/path/action")
.IfBody(s => s.ReadAsString().Contains("foo") && s.ReadAsString().Contains("bar"));
.Response((request, args) => "foobar response");